Updates On DatabaseManager class, UserDetails class and EventDetails class
added the ability to add UserDetails class obj and EventDetails class obj to be Firestore based on the current signed in google acc.
Updated all the getters and setters in the UserDetails and EventDetails class so that it can be used with the Google FireStore.
Added Debug functions to the above as well.